refactor: merge pull request tools into github focus
