+++ tests/pi/agent/extensions/user/lib/focus/definitions.test.ts
@@ -66,9 +66,9 @@ describe("focus definitions", () => {
 		assert.equal(isFocusTransition(undefined), false);
 	});
 
-	it("defines the pull-request focus as manual with PR workflow tools", () => {
-		const focus = BASE_FOCUS_DEFINITIONS.find((f) => f.name === "pull-request");
-		assert.ok(focus, "pull-request focus should be defined");
+	it("defines the github focus as manual with PR workflow tools", () => {
+		const focus = BASE_FOCUS_DEFINITIONS.find((f) => f.name === "github");
+		assert.ok(focus, "github focus should be defined");
 		assert.equal(focus?.transition, FOCUS_TRANSITION.CONFIRM);
 		assert.deepEqual(focus?.tools, [
 			"create_pull_request",
